+++ include/sys/ino.h
@@ -5,15 +5,15 @@
  * a disk block.
  */
 struct dinode {
-    unsigned short di_mode; /* mode and type of file */
-    short di_nlink;         /* number of links to file */
-    short di_uid;           /* owner's user id */
-    short di_gid;           /* owner's group id */
-    off_t di_size;          /* number of bytes in file */
-    char di_addr[40];       /* disk block addresses */
-    time_t di_atime;        /* time last accessed */
-    time_t di_mtime;        /* time last modified */
-    time_t di_ctime;        /* time created */
+    int di_mode;      /* mode and type of file */
+    int di_nlink;     /* number of links to file */
+    int di_uid;       /* owner's user id */
+    int di_gid;       /* owner's group id */
+    off_t di_size;    /* number of bytes in file */
+    char di_addr[40]; /* disk block addresses */
+    time_t di_atime;  /* time last accessed */
+    time_t di_mtime;  /* time last modified */
+    time_t di_ctime;  /* time created */
 };
 #define INOPB 8 /* 8 inodes per block */
 /*
